refactor for WES encapsulation, simplified graphql, have runs working
