+++ src/main/java/org/icgc_argo/workflow/search/model/graphql/Run.java
@@ -1,0 +1,4 @@
+package org.icgc_argo.workflow.search.model.graphql;
+
+public class Run {
+}
