+++ src/main/java/org/icgc_argo/workflow/search/model/wes/ServiceInfo.java
