+++ src/main/java/org/icgc_argo/workflow/search/repository/RunRepository.java
@@ -1,0 +1,4 @@
+package org.icgc_argo.workflow.search.repository;
+
+public class RunRepository {
+}
