+++ src/main/java/org/icgc_argo/workflow/search/service/WorkflowService.java
@@ -1,0 +1,4 @@
+package org.icgc_argo.workflow.search.service;
+
+public class WorkflowService {
+}
